+++ vsc-base/out/vsc-base-vscode.d.ts
@@ -1,5 +1,5 @@
-import * as vscode from 'vscode';
 import * as vsc from './vsc-base';
+import * as vscode from 'vscode';
 /** vsc-base method
  * @description
  * Prompt user for a question
@@ -228,7 +228,7 @@ export declare const prependToDocument: (content: string, editor?: vscode.TextEd
  * @description
  * Insert content at vscode.Range
  * Return true on success, false if the document or textEditor was not open/correct
- * @see [appendToDocument](http://vsc-base.org/#appendToDocument)
+ * @see [insertAtRange](http://vsc-base.org/#insertAtRange)
  * @param content
  * @param range
  * @param editor
@@ -242,38 +242,37 @@ export declare const insertAtRange: (content: string, range: vscode.Range, edito
  * @description
  * Insert content at position (start and optional end postion)
  * Return true on success, false if the document or textEditor was not open/correct
- * @see [appendToDocument](http://vsc-base.org/#appendToDocument)
+ * @see [insertAt](http://vsc-base.org/#insertAt)
  * @param content
  * @param range
  * @param editor
  * @dependencyExternal vscode
  * @vscType Vscode
- * @oneLineEx const success = await vsc.insertAtRange(content, range)
+ * @oneLineEx const success = await vsc.insertAt(content, start, end)
  * @returns Promise<boolean>
  */
 export declare const insertAt: (content: string, start: number, end?: number, editor?: vscode.TextEditor | undefined, trimSpaces?: boolean) => Promise<boolean>;
 /** vsc-base method
  * @description
  * Append new line content in the end of the (open) document
- * @see [appendLineToActiveDocument](http://vsc-base.org/#appendLineToActiveDocument)
+ * @see [appendLineToDocument](http://vsc-base.org/#appendLineToDocument)
  * @param content
  * @param editor
  * @dependencyInternal appendToActiveDocument
  * @vscType Vscode
- * @oneLineEx const success = await vsc.appendLineToActiveDocument(content)
+ * @oneLineEx const success = await vsc.appendLineToDocument(content)
  * @returns Promise<boolean>
  */
 export declare const appendLineToDocument: (content: string, editor?: vscode.TextEditor | undefined) => Promise<boolean>;
 /** vsc-base method
  * @description
  * Prepend new line content in the start of the (open) document
- * @see [appendLineToActiveDocument](http://vsc-base.org/#appendLineToActiveDocument)
+ * @see [prependLineToDocument](http://vsc-base.org/#prependLineToDocument)
  * @param content
  * @param document
  * @param editor
- * @dependencyInternal appendToActiveDocument
  * @vscType Vscode
- * @oneLineEx const success = await vsc.appendLineToActiveDocument(content)
+ * @oneLineEx const success = await vsc.prependLineToDocument(content)
  * @returns Promise<boolean>
  */
 export declare const prependLineToDocument: (content: string, editor?: vscode.TextEditor | undefined) => Promise<boolean>;
@@ -281,21 +280,21 @@ export declare const prependLineToDocument: (content: string, editor?: vscode.Te
  * @description
  * Save active open file. \
  * Return true for succes, and false if there was no open document
- * @see [saveActiveDocument](http://vsc-base.org/#saveActiveDocument)
+ * @see [saveDocument](http://vsc-base.org/#saveDocument)
  * @dependencyInternal getActiveDocument
  * @vscType Vscode
- * @oneLineEx const success = await vsc.saveActiveDocument(content)
+ * @oneLineEx const success = await vsc.saveDocument(content)
  * @returns Promise<boolean>
  */
 export declare const saveDocument: (document?: vscode.TextDocument | undefined) => Promise<boolean>;
 /** vsc-base method
  * @description
  * Takes a start and end and return vscode positons and range objects.
- * @see [getComplexRangeObject](http://vsc-base.org/#getComplexRangeObject)
+ * @see [createVscodeRangeAndPosition](http://vsc-base.org/#createVscodeRangeAndPosition)
  * @param range
  * @param editor
  * @vscType Vscode
- * @oneLineEx const success = vsc.getComplexRangeObject(source, start, end)
+ * @oneLineEx const success = vsc.createVscodeRangeAndPosition(source, start, end)
  * @returns boolean
  */
 export declare const createVscodeRangeAndPosition: (source: string, start: number, end?: number, trimSpaces?: boolean) => vsc.VscodePosition;
@@ -377,7 +376,7 @@ export declare const setSelectionFromRange: (range: vscode.Range, editor?: vscod
  * Set Selections for an TextEditor (Current document) \
  * Clear other selections \
  * returns true on success
- * @see [setSelectionsFromRanges](http://vsc-base.org/#setSelectionFromRange)
+ * @see [setSelectionsFromRanges](http://vsc-base.org/#setSelectionsFromRanges)
  * @param range
  * @param editor
  * @vscType Vscode