+++ vsc-script/.vsc-script/test.vsc-script.ts
@@ -2,5 +2,6 @@ import * as vsc from 'vsc-base'
 
 export async function run(path: string) {
    vsc.showMessage('Im a Test!')
-
 }
+
+